+++ butler_offline_selenium_tests/gemeinsame_buchungen_test.py
@@ -1,7 +1,5 @@
 from butler_offline_selenium_tests.SeleniumTest import SeleniumTestClass
-from butler_offline_selenium_tests.SeleniumTest import enter_test_mode
-from butler_offline_selenium_tests.SeleniumTest import define_kategorie
-from butler_offline_selenium_tests.SeleniumTest import get_selected_option
+from butler_offline_selenium_tests.page.util import enter_test_mode, define_kategorie, get_selected_option
 from butler_offline_selenium_tests.page.gemeinsam.gemeinsam_add import GemeinsamAdd
 from butler_offline_selenium_tests.page.gemeinsam.gemeinsam_uebersicht import GemeinsamUeberischt
 from butler_offline_selenium_tests.page.core.configuration import Configuration
