+++ butler_offline_selenium_tests/page/einzelbuchungen/einzelbuchung_add.py
@@ -1,5 +1,4 @@
-from butler_offline_selenium_tests.SeleniumTest import fill_element, select_option, click_add_button, \
-    get_selected_option
+from butler_offline_selenium_tests.page.util import fill_element, select_option, click_add_button, get_selected_option
 from selenium.webdriver.common.by import By
