+++ butler_offline_selenium_tests/page/sparen/depotwert_add.py
@@ -1,4 +1,4 @@
-from butler_offline_selenium_tests.SeleniumTest import fill_element, select_option, click_add_button
+from butler_offline_selenium_tests.page.util import fill_element, select_option, click_add_button
 
 
 class DepotwertAdd:
